refactor: delete dead LLM-routing scaffolding
Closes #86.

Three parallel code paths looked like they implemented features but
none were wired into production. Removing them so the code matches the
docs PR #85 just landed. Also picks up incidental ruff cleanups on the
files touched.

1. `_classify_via_ollama` removed from `api/services/chat_helpers.py`.
   The live intent classifier is `_classify_via_haiku` (with regex
   fallback); the Ollama variant was only mocked in tests.

2. `api/services/model_selector.py` deleted entirely — `CLAUDE_MODELS`,
   `get_claude_model_name`, `classify_query_complexity`, `ModelSelector`,
   `get_model_selector`, `reset_model_selector`. The agent loop
   (`agent_loop.py:91`) and the synthesizer (`synthesizer.py:116`) both
   document `model_tier` as ignored, so the tier classifier never
   actually changed the model. `tests/test_model_selector.py` deleted;
   `tests/reset_singletons.py` no longer resets the gone-singleton.
   `api/routes/chat.py` drops the `model_select` perf span and now
   labels the routing event with `LIFEOS_ANTHROPIC_MODEL` directly.
   `api/services/query_router.py` drops `recommended_model` /
   `complexity_score` fields on `RoutingResult` (no external readers)
   and stops calling `classify_query_complexity`.

3. `lifeos-llm.service` branch removed from `_stop_llm_for_embeddings`
   and `_start_llm` in `scripts/run_all_syncs.py`. The service is
   disabled on production and `lifeos-api.service` no longer `Wants=`
   it. The ollama unload/warm logic (PR #83) is what actually runs;
   keeping that intact. Also drops the `_lifeos_llm_was_stopped_by_us`
   flag and simplifies `_is_llm_running` to just check ollama VRAM.

4. e2e tests in `tests/e2e/test_telegram_*.py` retargeted from
   `_classify_via_ollama` to `_classify_via_haiku`. They were mocking
   the dead path, so the tests were effectively making real Anthropic
   calls — now they actually mock the live classifier.

5. README + AGENTS.md state that `LIFEOS_ANTHROPIC_MODEL` is the single
   orchestrator-model knob with no per-query tiering.

The sudoers `NOPASSWD` entry for `systemctl start/stop lifeos-llm.service`
on the production machine can be removed manually — this PR doesn't
touch sudoers since that's a per-installation concern.

Tests: ./scripts/test.sh — 1928 passed, 2 pre-existing data-integrity
failures unrelated to this change (test_top_person_stats_accurate,
test_all_vault_files_exist).
